azblobregex/pyproject.toml
[tool.poetry]
name = "azblobregex"
version = "0.1.0"
description = ""
authors = ["Niten Nashiki <niten.nashiki@lightz-inc.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.4"
azure-storage-blob = "^12.9.0"
tdqm = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
azblobregex = "azblobregex.command_line:cli"
